+++ pluto-plugins/plugins/datastore/lib/src/main/java/com/pluto/plugins/datastore/pref/internal/Preferences.kt
@@ -1,4 +1,4 @@
-package com.pluto.plugins.datastore.pref.utils
+package com.pluto.plugins.datastore.pref.internal
 
 import android.content.Context
