+++ pluto-plugins/plugins/datastore/lib/src/main/java/com/pluto/plugins/datastore/pref/internal/ui/DatastorePrefAdapter.kt
@@ -1,7 +1,7 @@
-package com.pluto.plugins.datastore.pref.ui
+package com.pluto.plugins.datastore.pref.internal.ui
 
 import android.view.ViewGroup
-import com.pluto.plugins.datastore.pref.utils.DatastorePrefKeyValuePair
+import com.pluto.plugins.datastore.pref.internal.DatastorePrefKeyValuePair
 import com.pluto.utilities.list.BaseAdapter
 import com.pluto.utilities.list.DiffAwareHolder
 import com.pluto.utilities.list.ListItem
