+++ pluto-plugins/plugins/datastore/lib/src/main/res/navigation/pluto_dts___navigation.xml
@@ -7,7 +7,7 @@
 
     <fragment
         android:id="@+id/listView"
-        android:name="com.pluto.plugins.datastore.pref.ui.ListFragment"
+        android:name="com.pluto.plugins.datastore.pref.internal.ui.ListFragment"
         android:label="list"
         tools:layout="@layout/pluto_dts___fragment_list">
     </fragment>
